segmented control 을 추가하여 main activity 안에 두 개의 fragment로 나누었음.
* main activity 안에 calendar fragment/ list fragment를 추가하여 toggle 확인이 가능하게 변경하였음
