+++ app/src/main/java/com/aday/abook/ApplicationComponent.kt
@@ -2,11 +2,12 @@ package com.aday.abook
 
 import android.app.Application
 import android.app.IntentService
-import androidx.room.Room
-import com.aday.abook.feature.calendar.CalendarComponent
 import com.aday.abook.feature.main.MainComponent
+import com.aday.abook.feature.addbook.AddBookComponent
 import com.aday.abook.feature.memo.BookMemoComponent
 import com.aday.abook.feature.search.BookSearchComponent
+import com.aday.abook.fragment.calendar.CalendarComponent
+import com.aday.abook.fragment.list.BookListComponent
 import com.aday.abook.mvvm.ViewModelModule
 import com.aday.core.dagger.module.APIModule
 import com.aday.core.dagger.module.ApplicationModule
@@ -39,12 +40,16 @@ interface ApplicationComponent {
     fun inject(intentService: IntentService)
 
     //add component builder 여기로 들어오면 이제 다른 컨포넌트로 들어갈 수 있음
-    fun mainComponentBuilder(): MainComponent.Builder
+    fun addBookComponentBuilder(): AddBookComponent.Builder
 
     fun bookMemoComponentBuilder(): BookMemoComponent.Builder
 
     fun bookSearchComponentBuilder(): BookSearchComponent.Builder
 
+    fun mainComponentBuilder(): MainComponent.Builder
+
     fun calendarComponentBuilder(): CalendarComponent.Builder
+
+    fun bookListComponentBuilder(): BookListComponent.Builder
 
 }
