+++ app/src/main/res/layout/fragment_calendar_view.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout>
+    <data>
+        <variable
+            name="fragmentViewModel"
+            type="com.aday.abook.feature.main.MainViewModel" />
+    </data>
+<com.prolificinteractive.materialcalendarview.MaterialCalendarView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/calendarView"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="8dp"
+    app:layout_constraintStart_toStartOf="parent"
+    app:mcv_selectionColor="@color/brown1"
+    app:mcv_showOtherDates="all" />
+</layout>
+
