+++ app/src/main/res/values/colors.xml
@@ -15,4 +15,16 @@
     <color name="brown3">#A69C90</color>
 
     <color name="grey1">#80C2C2C2</color>
+
+    <color name="default_text_color">#3e3e3d</color>
+    <color name="default_text_color_checked">#ffffff</color>
+
+    <color name="default_segment_color">#f8f8f8</color>
+    <color name="default_segment_color_checked">#7f6a93</color>
+
+    <color name="default_ripple_color">#e4e4e4</color>
+    <color name="default_ripple_color_checked">#c695a4</color>
+
+    <color name="default_border_color">#e4e4e4</color>
+    
 </resources>
