+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="compound_drawable_size">24dp</dimen>
+    <dimen name="default_segment_text_size">16sp</dimen>
+    <dimen name="default_segment_height">40dp</dimen>
+    <dimen name="default_border_width">1dp</dimen>
+    <dimen name="default_corner_radius">8dp</dimen>
+
+</resources>
