+++ app/src/main/res/values/styles.xml
@@ -23,4 +23,5 @@
     <item name="android:windowFullscreen">true</item>
     <item name="android:fontFamily">@font/font</item>
     </style>
+
 </resources>
